More poller work and the start of plugins
Reworked a big chunk of the poller logic for the eventual support
of backends other than sensu. Added crude plugin support that kind
of works. Ditched aiohttp for requests, but running it within a
coroutine so the performance should be about the same. I kept
hitting weird bugs with aiohttp that got super old super fast.
Started to change the naming convention from 'datacenters' to
'sources' since that makes more sense. Renamed it in the config
file, will rename throughout the rest of the app later. Pollers
are now responsible for a single source (so one for region1, one for
region2) and I added a scheduler class that handles firing off the
pollers. There are two scheduler classes at the moment, one for
uwsgi and one for running stand-alone. The stand-alone class is
super crappy at the moment and not really needed for running in
prod, so the code is there, but nothing touches it. The uwsgi
scheduler works well, but that's not surprising considering that
uwsgi does all the work, lol.

Signed-off-by: M. David Bennett <mdavidbennett@syntheticworks.com>