Moving zone and dashboard config to the DB (#208)
Moved zone and dashboard definitions into the DB so they can be dynamically created in the future. This will also pave the way to user-created and owned dashboards. Switched API layer over to flask-restful. This will also be done elsewhere in the code once user-auth is added. Made a bunch of convenience changes to the base cache models so they're easier to work with. Also added an id field to pretty much everything so zones and dashboards can be renamed on the fly.

Signed-off-by: M. David Bennett <mdavidbennett@syntheticworks.com>