SRE-4659 - Add DB Migration Support (#5)
This adds flask-migrate to Opsy and includes an initial schema. This
also introduces a new set of commands on the shell for db manipulation
under the "db" parent command. The command "init-db" has been moved
under the "db" parent command. The readme has been updated to reflect
this change and to add some docs about working with flask-migrate within
Opsy.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>