Auth routes (#240)
* Add query and ref schemas for user, role, and rolepermission

* fixing change to nested call

* spelling

* convert missed ma.Nested's

* Python 3.7 proofing

* Field revisions

* linting fixes

* Sre 5036 more routes (#7)

* WIP commit

* role permission views 1st pass

* Add Update Schema

Add update schema for all models. This makes it so that fields that are
marked as required for object creations aren't required for object
updates (things like "name").

Also fixed some minor bugs I found along the way.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Auth routes and supporting stuff

Renamed the opsy command to opsyctl so it no longer conflicts with the
new opsyclient. Updated the readme to reflect this name change.

Finished adding auth routes and removed all the auth-related CLI commands.
Added route for patch on login for the logged in user to perform
self-service on their account, things like changing password or email.
Removed prettytables as a requirement since it was mostly used for the
auth CLI stuff. Converted everything else that used it to just dumping
JSON. Added a new create-admin-user command to the CLI for creating
the initial local admin account. Updated readme with new getting started
instructions to support this change.

Updated the login logic so it will still allow local accounts when LDAP
is enabled. This is needed for the local admin user to be able to configure
LDAP roles. Added a new column to the user model to indicate if it should
be logged in through LDAP. Added supporting DB schema migration scripts for
this change.

Still need tests for these changes, will address in the next commit.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Fix timezones because they're serializing wrong

Apparently the latest version of marshmallow changed how it works with
timezone naive datetime objects. This just makes all timestamps default
to being UTC.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Unit tests for auth routes

Also fixed bugs along the way.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Re-pin apispec to 2.0.2

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>