Add version command and fix bug with logging
Added a command to opsyctl to print the version. Fixed the value of
opsy.__version__ to include VCS status, it should only show a full
version if this is a tagged release.

Updated opsy.utils.print_error to be able to print exceptions. Added
tests for this.

Fixed a bug in logging where an extra handler would log messages.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>